+++ src/NServiceBus.NHibernate.AcceptanceTests/TestSuiteConstraints.cs
@@ -10,7 +10,7 @@
         public bool SupportsDelayedDelivery => true;
         public bool SupportsOutbox => true;
         public bool SupportsPurgeOnStartup => true;
-        public IConfigureEndpointTestExecution CreateTransportConfiguration() => new ConfigureEndpointAcceptanceTestingTransport(SupportsNativePubSub, SupportsDelayedDelivery);
+        public IConfigureEndpointTestExecution CreateTransportConfiguration() => new ConfigureEndpointAcceptanceTestingTransport(SupportsNativePubSub, SupportsDelayedDelivery, TransportTransactionMode.ReceiveOnly);
         public IConfigureEndpointTestExecution CreatePersistenceConfiguration() => new ConfigureEndpointNHibernatePersistence();
     }
 }
