+++ src/NServiceBus.NHibernate.PersistenceTests/NServiceBus.NHibernate.PersistenceTests.csproj
@@ -14,7 +14,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
     <PackageReference Include="NHibernate" Version="5.3.11" />
-    <PackageReference Include="NServiceBus.PersistenceTests.Sources" Version="8.0.0-beta.1" GeneratePathProperty="true" />
+    <PackageReference Include="NServiceBus.PersistenceTests.Sources" Version="8.0.0-storagession.2" GeneratePathProperty="true" />
     <PackageReference Include="NUnit" Version="3.13.3" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
     <PackageReference Include="System.Data.SqlClient" Version="4.8.3" />
