+++ src/NServiceBus.NHibernate/NServiceBus.NHibernate.csproj
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup Label="Public dependencies">
-    <PackageReference Include="NServiceBus" Version="[8.0.0-beta.1, 9.0.0)" />
+    <PackageReference Include="NServiceBus" Version="8.0.0-storagession.2" />
     <PackageReference Include="NHibernate" Version="[5.3.0, 6.0.0)" />
   </ItemGroup>
