+++ src/NServiceBus.NHibernate/SynchronizedStorage/SessionFactoryHolder.cs
@@ -1,0 +1,14 @@
+namespace NServiceBus.NHibernate.SynchronizedStorage
+{
+    using global::NHibernate;
+
+    class SessionFactoryHolder
+    {
+        public SessionFactoryHolder(ISessionFactory sessionFactory)
+        {
+            SessionFactory = sessionFactory;
+        }
+
+        public ISessionFactory SessionFactory { get; }
+    }
+}
