--- src/NServiceBus.NHibernate/SynchronizedStorage/SharedMappings.cs
@@ -1,20 +1,0 @@
-namespace NServiceBus.Features
-{
-    using System;
-    using global::NHibernate.Cfg;
-
-    class SharedMappings
-    {
-        public void AddMapping(Action<Configuration> mapping)
-        {
-            mappings.Add(mapping);
-        }
-
-        public void ApplyTo(Configuration configuration)
-        {
-            mappings.ForEach(m => m(configuration));
-        }
-
-        System.Collections.Generic.List<Action<Configuration>> mappings = new System.Collections.Generic.List<Action<Configuration>>();
-    }
-}
